implement room service. It's working but not quite. TODO: sounds, players hit test

+++ core/src/main/kotlin/com/catinthedark/models/Model.kt
@@ -29,7 +29,7 @@ data class BrickModel(
     val id: UUID,
     var x: Float,
     var y: Float,
-    var angle: Float,
+    var angle: Double,
     var hurting: Boolean) : Model
 
 data class BonusModel(
